front: updateTrain: rename & passing AddOrEditTrainSchedule to TSX
