+++ front/src/applications/operationalStudies/components/ManageTrainSchedule/AddOrEditTrainSchedule/trainNameHelper.jsx
