--- front/src/applications/operationalStudies/components/ManageTrainSchedule/AddTrainLabels.jsx
@@ -1,32 +1,0 @@
-import React from 'react';
-import { useTranslation } from 'react-i18next';
-import { useSelector, useDispatch } from 'react-redux';
-import { updateLabels } from 'reducers/osrdconf';
-import ChipsSNCF from 'common/BootstrapSNCF/ChipsSNCF';
-import { getLabels } from 'reducers/osrdconf/selectors';
-
-export default function AddTrainLabels() {
-  const { t } = useTranslation(['operationalStudies/manageTrainSchedule']);
-  const labels = useSelector(getLabels);
-  const dispatch = useDispatch();
-
-  const removeTag = (idx) => {
-    const newTags = Array.from(labels);
-    newTags.splice(idx, 1);
-    dispatch(updateLabels(newTags));
-  };
-
-  const addTag = (tag) => {
-    const newTags = Array.from(labels);
-    newTags.push(tag);
-    dispatch(updateLabels(newTags));
-  };
-
-  return (
-    <div className="osrd-config-item" data-testid="add-train-labels">
-      <div className="osrd-config-item-container mb-2">
-        <ChipsSNCF addTag={addTag} tags={labels} removeTag={removeTag} title={t('trainLabels')} />
-      </div>
-    </div>
-  );
-}
