+++ front/src/reducers/osrdsimulation/simulation.js
@@ -6,7 +6,7 @@ import {
 import { deleteRequest, get, post } from 'common/requests';
 import { setFailure } from 'reducers/main';
 
-import trainNameWithNum from 'applications/operationalStudies/components/ManageTrainSchedule/AddTrainSchedule/trainNameHelper';
+import trainNameWithNum from 'applications/operationalStudies/components/ManageTrainSchedule/AddOrEditTrainSchedule/trainNameHelper';
 import { trainscheduleURI } from 'applications/operationalStudies/components/SimulationResults/simulationResultsConsts';
 
 export const UPDATE_SIMULATION = 'osrdsimu/UPDATE_SIMULATION';
